+++ src/main/scala/com/github/gchudnov/mtg/Diagram.scala
@@ -2,6 +2,7 @@ package com.github.gchudnov.mtg
 
 import com.github.gchudnov.mtg.diagram.Renderer
 import com.github.gchudnov.mtg.diagram.Span
+import com.github.gchudnov.mtg.diagram.Theme
 import com.github.gchudnov.mtg.diagram.internal.BasicRenderer
 import com.github.gchudnov.mtg.diagram.internal.BasicTranslator
 import scala.collection.mutable.ArrayBuffer
@@ -26,6 +27,7 @@ final case class Diagram(
 object Diagram:
 
   export com.github.gchudnov.mtg.diagram.Span
+  export com.github.gchudnov.mtg.diagram.Theme
 
   val empty: Diagram =
     Diagram(
@@ -37,55 +39,6 @@ object Diagram:
       legends = List.empty[Legend],
       annotations = List.empty[Annotation]
     )
-
-  /**
-   * Theme
-   *
-   * Provides rendering options.
-   */
-  final case class Theme(
-    space: Char,
-    fill: Char,
-    leftOpen: Char,
-    leftClosed: Char,
-    rightOpen: Char,
-    rightClosed: Char,
-    axis: Char,
-    tick: Char,
-    border: Char,
-    comment: Char,
-    legend: Boolean,
-    annotations: Boolean,
-    label: Theme.Label
-  ):
-    def leftBound(isInclude: Boolean): Char =
-      if isInclude then leftClosed else leftOpen
-
-    def rightBound(isInclude: Boolean): Char =
-      if isInclude then rightClosed else rightOpen
-
-  object Theme:
-    enum Label:
-      case None      // draw labels as-is on one line
-      case NoOverlap // draw sorted labels that are non-overlapping, some of the labels might be skipped
-      case Stacked   // draw all labels, but stack them onto multiple lines
-
-    val default: Theme =
-      Theme(
-        space = ' ',
-        fill = '*',
-        leftOpen = '(',
-        leftClosed = '[',
-        rightOpen = ')',
-        rightClosed = ']',
-        axis = '-',
-        tick = '+',
-        border = '|',
-        comment = ':',
-        legend = true,
-        annotations = true,
-        label = Label.NoOverlap
-      )
 
   /**
    * View
