+++ src/main/scala/com/github/gchudnov/mtg/diagram/Span.scala
@@ -5,7 +5,11 @@ import com.github.gchudnov.mtg.Diagram.Tick
 /**
  * Span
  *
- * NOTE: Use negative span for an empty interval
+ * Used to draw an interval on the canvas.
+ *
+ *   - NOTE: Use negative span for an empty interval
+ *   - NOTE: The span is inclusive on both ends.
+ *   - NOTE: includeX0 and includeX1 are used to draw begin and end of the span and do not affect the size of the span.
  */
 final case class Span(x0: Int, x1: Int, includeX0: Boolean, includeX1: Boolean):
